+++ packages/frontend/src/components/Settings/Privacy.tsx
@@ -8,6 +8,8 @@ import useTranslationFunction from '../../hooks/useTranslationFunction'
 import useDialog from '../../hooks/dialog/useDialog'
 import UnblockContacts from '../dialogs/UnblockContacts'
 import SettingsButton from './SettingsButton'
+import DesktopSettingsSwitch from './DesktopSettingsSwitch'
+import { runtime } from '@deltachat-desktop/runtime-interface'
 
 import type { DesktopSettingsType } from '../../../../shared/shared-types'
 import type { SettingsStoreState } from '../../stores/settings'
@@ -31,6 +33,13 @@ export default function Privacy({ settingsStore, desktopSettings }: Props) {
         label={tx('pref_read_receipts')}
         description={tx('pref_read_receipts_explain')}
       />
+      {runtime.getRuntimeInfo().isContentProtectionSupported && (
+        <DesktopSettingsSwitch
+          settingsKey='contentProtectionEnabled'
+          label={tx('pref_screen_security')}
+          description={tx('pref_screen_security_explain')}
+        />
+      )}
       <SettingsSeparator />
       <SettingsHeading>{tx('delete_old_messages')}</SettingsHeading>
       <Autodelete settingsStore={settingsStore} />